refactor, renaming Player component (& route) to NowPlaying
