+++ imports/ui/containers/tests/__stories__/App.stories.jsx
@@ -6,7 +6,7 @@ import {
 } from "@kadira/storybook";
 import App from "../../App";
 import Welcome from "../../../components/Welcome";
-import Player from "../../../components/Player";
+import NowPlaying from "../../../components/NowPlaying";
 
 storiesOf("Pages", module)
   .add("App - Welcome", (): React$Element<*> => (
@@ -14,8 +14,8 @@ storiesOf("Pages", module)
       <Welcome />
     </App>
   ))
-  .add("App - Player", (): React$Element<*> => (
+  .add("App - Now Playing", (): React$Element<*> => (
     <App>
-      <Player />
+      <NowPlaying />
     </App>
   ));
